MVP/backend/requirements.txt
fastapi
uvicorn
google-generativeai
python-multipart
PyPDF2
python-dotenv
